openkylin_agent/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
openai==2.3.0
pydantic==2.12.0
pydantic_core==2.41.1
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
